product-read-kafka-consumers/requirements.txt

kafka-python
sqlalchemy
psycopg2-binary
